12/19 추가 수정 반영 -- 알파테스트 마무리

+++ frontend/src/components/mypage/Navbar.js
@@ -4,14 +4,15 @@ import styles from '../../scss/mypage/Navbar.module.scss';
 
 function Navbar() {
     const navigate = useNavigate();
+    const customerName = sessionStorage.getItem('customerName'); // 세션에서 이름 가져오기
 
-    const Logout = () => {
-        const customerName = sessionStorage.getItem('customerName'); // 세션에서 이름 가져오기
-        alert(`${customerName}님 로그아웃되었습니다.`);
-
-        sessionStorage.clear(); // 세션 스토리지 초기화
-
-        navigate('/');
+    const Logout = e => {
+        e.preventDefault(); // Link의 기본 이동 동작 방지
+        const logoutChk = window.confirm(`${customerName}님, 로그아웃 하시겠습니까?`);
+        if (logoutChk) {
+            sessionStorage.clear(); // 세션 스토리지 초기화
+            navigate('/'); // 홈으로 이동
+        }
     };
 
     return (
